requirements.txt
# Prototipo_chatbot - Dependencies COMPLETAS

# ===== CORE FRAMEWORK =====
flask>=2.3.0
pydantic>=2.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.0

# ===== AI & ML =====
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0

# ===== VECTOR DATABASES =====
faiss-cpu>=1.7.4
chromadb>=0.4.15

# ===== LOCAL MODELS =====
openai>=1.3.0

# ===== DOCUMENT PROCESSING =====
PyPDF2>=3.0.1
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ===== WEB SCRAPING =====
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# ===== DATABASE =====
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# ===== LOGGING & MONITORING =====
structlog>=23.0.0
colorlog>=6.7.0

# ===== DEVELOPMENT =====
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# ===== DEPLOYMENT =====
gunicorn>=21.2.0